requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
apscheduler==3.10.4
google-genai>=1.0.0
pydantic-settings==2.7.0
pandas>=2.2.3
pandas_ta>=0.4.71b0
numpy>=2.0.2
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
xgboost==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
pyarrow==17.0.0
google-cloud-storage==2.18.2
